+++ projects/plugins/jetpack/extensions/plugins/image-studio/image-studio.php
@@ -420,14 +420,10 @@ function get_tracking_site_type() {
  * @return bool True when the current visitor is an Automattician.
  */
 function is_tracking_automattician() {
-	if ( function_exists( 'wpcom_is_proxied_request' )
-		&& \wpcom_is_proxied_request()
-		&& function_exists( 'is_automattician' )
-	) {
-		return (bool) \is_automattician();
-	}
+	$is_automattician = function_exists( 'is_automattician' ) && (bool) \is_automattician();
+	$is_wpcom_proxied = function_exists( 'wpcom_is_proxied_request' ) && \wpcom_is_proxied_request();
 
-	return ( new Visitor() )->is_automattician_feature_flags_only();
+	return $is_automattician || $is_wpcom_proxied || ( new Visitor() )->is_automattician_feature_flags_only();
 }
 
 /**